+++ R/utils-exported.R
@@ -567,6 +567,7 @@ getBGMSMSPeaks <- function(anaInfo, replicates = NULL, MSLevel = 2, retentionRan
                                 startMobs = mobilityRange[1], endMobs = mobilityRange[2],
                                 relMinAbundance = avgSpectraParams$relMinAbundance,
                                 absMinAbundance = avgSpectraParams$absMinAbundance,
+                                minRelCumIntensity = avgSpectraParams$minRelCumIntensity,
                                 smoothWindowIMS = avgSpectraParams$smoothWindowIMS,
                                 halfWindowIMS = avgSpectraParams$halfWindowIMS,
                                 maxGapIMS = avgSpectraParams$maxGapIMS,
@@ -585,7 +586,8 @@ getBGMSMSPeaks <- function(anaInfo, replicates = NULL, MSLevel = 2, retentionRan
     ret <- averageSpectraList(list(blSpecs), avgAnalysesParams$clusterMzWindow, avgAnalysesParams$topMost,
                               avgAnalysesParams$minIntensityPre, avgAnalysesParams$minIntensityPost,
                               avgAnalysesParams$relMinAbundance, avgAnalysesParams$absMinAbundance,
-                              avgAnalysesParams$method, FALSE, FALSE, FALSE, FALSE)[[1]]
+                              avgAnalysesParams$minRelCumIntensity, avgAnalysesParams$method, FALSE, FALSE, FALSE,
+                              FALSE)[[1]]
     ret[, precursor := NULL]
     setnames(ret, c("abundance_rel", "abundance_abs", "abundance_prev_rel", "abundance_prev_abs"),
              c("abundance_rel_ana", "abundance_abs_ana", "abundance_rel_spec", "abundance_abs_spec"))